update: make lag_matrix dtype int in FunctionalConnectivityLaggedBase

+++ junifer/markers/functional_connectivity/functional_connectivity_lagged_base.py
@@ -131,7 +131,7 @@ class FunctionalConnectivityLaggedBase(BaseMarker):
         fmri_data = aggregation["aggregation"]["data"].T
         n_rois = fmri_data.shape[0]
         fc_matrix = np.zeros((n_rois, n_rois))
-        lag_matrix = np.zeros((n_rois, n_rois))
+        lag_matrix = np.zeros((n_rois, n_rois), dtype=int)
         lags = np.arange(-self.max_lag, self.max_lag + 1)
         # Compute
         for i, j in product(range(n_rois), range(n_rois)):
